Adding un/re-subscribe test scenario placeholders

+++ src/main/java/org/eaa690/aerie/model/wx/Dewpoint.java
@@ -38,11 +38,11 @@ public class Dewpoint implements Serializable {
     /**
      * Celsius.
      */
-    private Integer celsius;
+    private Long celsius;
 
     /**
      * Fahrenheit.
      */
-    private Integer fahrenheit;
+    private Long fahrenheit;
 
 }
